// src/model/Social.ts
export type SocialSection = {
    instagram: string;
    facebook: string;
    x: string;
    tiktok: string;
    youtube: string;
    linkedin: string;
    pinterest: string;
    twitch: string;
    github: string;
    threads: string;
    snapchat: string;
    patreon: string;
    behance: string;
    dribbble: string;
    gitlab: string;
    codepen: string;
    stackoverflow: string;

    updatedAt: Date;
};